+++ crates/node/src/manager/node/close_epoch.rs
